recover even fec blocks and not just data
Had forgotten to recover fec blocks previously. Now it has been
corrected so that even fec blocks get recovered. This is required
because if there are 2 or more errors in a given row or col, then
even fec blocks will be required to recover from the error, only
data blocks won't help. And in turn if the fec block which is
required for error recovery of such a row or col is corrupted,
then unless that fec block is recovered, the data block wont be
recoverable.

TOCHECK: The random error generated currently is 16 bit long, while
the buf is a 8bit i.e a byte array, so if the lsbyte is 0 in the error,
then it won't/maynot get copied into the data/fec buffer in the
errorinjector logic.