block recovery implementation 1 of 2
Updating the fec_recover to actually recover the error blocks. So
have replaced the earlier simple block error flag clearing logic with
actual block recovery logic, which recovers the block and then inturn
clears the block error flag in fatMatrixFlag variable.

For now logic implemented to recover a given error block by using
the columns in the row where the error block resides. This is done
by calculating the xor along the columns in the row with error block
except for the error block itself, and inturn storing this calculated
value into the error block.

GENERAL NOTE (Valid for fec_recover logic from beginning): As of now
only the iContinue flag is used and iDone is not used, even thou both
are updated/calculated as part of the recovery process during each
iteration. Rather if we exit out of the recovery loop, it is possible
for two reasons. Reason1, we have recovered all error blocks so thus
iDone will also be set. Reason2, we have no more recoverable blocks,
maybe because there is groups of cyclic error blocks (4 blocks in
each such group), in which case iDone will not be set. So if iDone
is not set at end of fec_recover, then it means there are error
blocks, which require to be recovered using some other mechanism.