Makefile updated
to take care of the new test modes of the test program

removed depedence of clean on fec, as this was/will lead to
fec being created, when clean is called and fec is not already
present.

Also /bin/true added to rm commands so that the make rule wont
stop half way through due to one of the files being cleaned not
being present already.

Added logic to generate a random file as the data file for testing