WeakCrossCheck Between FineGrainedUserFlagging and LibraryGlobalCheck
A Weak cross check between user specified error blocks and what
library found at global level. At Global level using fec_checkfec
we can only find errors which don't complement one another out
in a cyclic manner.

Also we cann't tell if the error in a given row or col is due to
which specific error in the cols or rows within that row/col, if
more than 1 error.

Because of above reasons is why I am naming it as a WeakCrossCheck
Have to think of above once again later.