checkfec updated to check validity of data rows/cols only
As the fec row and col doesn't have fec data calculated and stored
for the fec row or col, this check is avoided now.

NOTE: cann't store fec data for both fec row as well as fec col,
as the fec of fec row/col will have to use the same location in the
fec matrix i.e (dmatrix+1, dmatrix+1) position. So fec can be
calculated / rather_stored only for the fec row or the fec col,
but not both. This has to be done at a later stage and handled
appropriately in weakcrosscheck and recovery logic also.